+++ flutter_record_platform_interface/lib/flutter_record_platform_interface.dart
@@ -1,2 +1,0 @@
-export 'package:flutter_record_platform_interface/src/flutter_record_platform_interface.dart';
-export 'package:flutter_record_platform_interface/src/types/types.dart';