+++ flutter_record_platform_interface/lib/src/types/types.dart
@@ -1,2 +1,0 @@
-export 'package:flutter_record_platform_interface/src/types/amplitude.dart';
-export 'package:flutter_record_platform_interface/src/types/audio_encoder.dart';