+++ flutter_sound_record_platform_interface/lib/flutter_sound_record_platform_interface.dart
@@ -1,0 +1,2 @@
+export 'package:flutter_sound_record_platform_interface/src/flutter_sound_record_platform_interface.dart';
+export 'package:flutter_sound_record_platform_interface/src/types/types.dart';