+++ flutter_sound_record_platform_interface/lib/src/types/types.dart
@@ -1,0 +1,2 @@
+export 'package:flutter_sound_record_platform_interface/src/types/amplitude.dart';
+export 'package:flutter_sound_record_platform_interface/src/types/audio_encoder.dart';