+++ ios/Classes/FlutterRecordPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterRecordPlugin : NSObject <FlutterPlugin>
-@end