+++ lib/flutter_record.dart
@@ -1,2 +1,0 @@
-export 'package:flutter_record/src/flutter_record.dart';
-export 'package:flutter_record_platform_interface/flutter_record_platform_interface.dart';