+++ lib/flutter_sound_record.dart
@@ -1,0 +1,2 @@
+export 'package:flutter_sound_record/src/flutter_sound_record.dart';
+export 'package:flutter_sound_record_platform_interface/flutter_sound_record_platform_interface.dart';